# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.SwaggerMessage:
    properties:
      details:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  ports.CreateEmployeeRequest:
    properties:
      createdAt:
        type: string
      department:
        type: string
      email:
        type: string
      employeeCode:
        type: string
      firstName:
        type: string
      hireDate:
        type: string
      hourlyRate:
        type: number
      hoursPerWeek:
        type: integer
      hoursWorked:
        type: number
      isActive:
        type: boolean
      lastName:
        type: string
      phoneNumber:
        type: string
      position:
        type: string
      role:
        type: string
      salary:
        type: number
      updatedAt:
        type: string
      userId:
        type: string
    required:
    - email
    - position
    type: object
...
  ports.UpdateEmployeeRequest:
    properties:
      department:
        type: string
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
      position:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'API de gestión de taller: autenticación JWT, coches, citas y empleados.'
  title: GonsGarage API
  version: "1.0"
paths:
...
  /api/v1/employees:
    get:
      parameters:
      - description: Límite (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Filtro departamento
        in: query
        name: department
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: employees, total, limit, offset
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Listar empleados
      tags:
      - employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del empleado
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ports.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: message y employee
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Crear empleado
      tags:
      - employees
  /api/v1/employees/{id}:
    delete:
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Eliminar empleado
      tags:
      - employees
    get:
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: employee
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Obtener empleado
      tags:
      - employees
    put:
      consumes:
      - application/json
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: string
      - description: Campos
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ports.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message y employee
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Actualizar empleado
      tags:
      - employees
...
securityDefinitions:
  BearerAuth:
    description: 'JWT: cabecera Authorization con valor Bearer seguido del token (rutas
      bajo /api/v1 salvo /health).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"